vtworker: Do not respond with a retryable error when the job was cancelled but not reset.
Since Cancel (or the completion of the current command) and Reset are no atomic operation, it's possible that a client might observe the following:
1. job running => client retries
2. job stopped, but not reset yet => client should retry
3. job is reset => client will succeed

To make sure that the client retries in 2. vtworker will respect a grace period of a minute between cancel/completion and reset. If the cancel/completion is more than one minute ago, the client in step 2 will see a non-retryable error.

The new filed, which preserves the time the job was cancelled/completed, will be shown in the status as well.