worker: Add suffix "Legacy" to VerticalSplitClone files.
This is in preparation for creating a new VerticalSplitClone which is based on the new online/offline SplitClone code.

We need to keep around the old VerticalSplit clone code to support textual primary key columns for the case where we have multiple source or destination shards and we need to merge them all.