test: Temporarily disable the webdriver test for the new vtctld2.
The test requires many dependencies which prolong the current test duration by up to 10 minutes on Travis.

Once we figured out how to cache the dependencies both on Travis and our local machines, we can re-enable the test.

Instead of commenting the code for installing the dependencies, I've moved it to bootstrap_web.sh for now.