throttler: Color the HTTP page based on the tested state and not the new state.
Somethings we test for a state, but skip the processing for different reasons. E.g. during "decrease" we might not have a previous value since the last rate change and without that value we cannot calculate the slave rate.

By coloring these results as well, it will be easier to debug the behavior of the code.