discovery: Add monitoring variable for the number of times a shard's master failed over and was promoted.
This variable will be used to verify that the vtgate buffer always detected a failover when the buffer stopped its buffering.

Note that the number of stopped bufferings can be lower. This is the case when:
a) a non-bufferable error happened (e.g. during COMMIT) and buffering was not started,
b) or no error occurred in vtgate (due to low QPS) and no buffering was started
c) or buffering was skipped because a reparent or failover was too recent.

a) is accounted for in the vtgate_buffer.py end to end test.

To track c), I've added the monitoring variable "BufferStartsSkipped".