vitess.io: Change URL format from "$page.html" to $page/.
This change only affects non-index pages e.g.
http://vitess.io/getting-started/local-instance.html.

To make sure that the previous URLs are still working, I've added
redirects for them.

This change also broke the internal links. Note that I did not fix them
and instead rewrote all of them with the "%{ link ... %}" tag. This has
the advantage that Jekyll now checks if the linked file exists.

Other changes:
- Updated "jekyll" and "liquid" versions. I'll rebuild the
"publish-site" Docker image shortly after this is merged.